Versao com relacionamentos

+++ src/main/java/rentvent/EnderecosRepo.java
@@ -2,6 +2,6 @@ package rentvent;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface EnderecosRepo extends CrudRepository<Enderecos, Integer>{
+public interface EnderecosRepo extends CrudRepository<Enderecos, Long>{
     
 }
