+++ src/main/java/rentvent/PessoaFisicaRepo.java
@@ -2,6 +2,6 @@ package rentvent;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PessoaFisicaRepo extends CrudRepository<PessoaFisica, Integer> {
+public interface PessoaFisicaRepo extends CrudRepository<PessoaFisica, Long> {
     
 }
